Move labels into a structured format
Split labels between object type e.g. car, human. Then human
is split into subcategories like face, hand, pose etc. This
helps a lot with organization and will allow us to expand more
easily.
