+++ menpo/landmark/labels/__init__.py
@@ -1,0 +1,6 @@
+from .base import labeller
+from .human import *
+from .car import (
+    streetscene_car_view_0, streetscene_car_view_1, streetscene_car_view_2,
+    streetscene_car_view_3, streetscene_car_view_4, streetscene_car_view_5,
+    streetscene_car_view_6, streetscene_car_view_7)
