+++ menpo/landmark/labels/human/__init__.py
@@ -1,0 +1,12 @@
+from .face import (
+    ibug_tongue,
+    ibug_close_eye_points, ibug_close_eye_trimesh, ibug_open_eye,
+    ibug_open_eye_trimesh,
+    ibug_face_49, ibug_face_49_trimesh, ibug_face_51, ibug_face_51_trimesh,
+    ibug_face_65_closed_mouth, ibug_face_66, ibug_face_66_trimesh,
+    ibug_face_68, ibug_face_68_trimesh,
+    imm_face, lfpw_face)
+from .hand import ibug_hand
+from .mesh import bu3dfe_83
+from .pose import (human36M_pose_17, human36M_pose_32, flic_pose, stickmen_pose,
+                   lsp_pose)
