+++ menpo/landmark/labels/human/face.py
@@ -1,139 +1,10 @@
 from collections import OrderedDict
 import numpy as np
-
 from menpo.landmark.base import LandmarkGroup
-from menpo.landmark.exceptions import LabellingError
 
-
-def _connectivity_from_array(array, close_loop=False):
-    r"""
-    Build the connectivity over a given array. For example, given ::
-
-        array = [(0, 3, 1, 2 )]
-
-    Generate the connectivity of ::
-
-        [(0, 3), (3, 1), (1, 2)]
-
-    If ``close_loop`` is true, add an extra connection from the last point to
-    the first.
-    """
-    # zip is a generator - need a list in this case
-    conn = list(zip(array, array[1:]))
-    if close_loop:
-        conn.append((array[-1], array[0]))
-    return np.asarray(conn)
-
-
-def _connectivity_from_range(range_tuple, close_loop=False):
-    r"""
-    Build the connectivity over a range. For example, given ::
-
-        range_array = np.arange(3)
-
-    Generate the connectivity of ::
-
-        [(0, 1), (1, 2), (2, 3)]
-
-    If ``close_loop`` is true, add an extra connection from the last point to
-    the first.
-    """
-    return _connectivity_from_array(
-        np.arange(*range_tuple), close_loop=close_loop)
-
-
-def _mask_from_range(range_tuple, n_points):
-    r"""
-    Generate a mask over the range. The mask will be true inside the range.
-    """
-    mask = np.zeros(n_points, dtype=np.bool)
-    range_slice = slice(*range_tuple)
-    mask[range_slice] = True
-    return mask
-
-
-def _build_labelling_error_msg(group, n_expected_points,
-                               n_actual_points):
-    return '{} mark-up expects exactly {} ' \
-           'points. However, the given landmark group ' \
-           'has {} points'.format(group, n_expected_points,
-                                  n_actual_points)
-
-
-def _validate_input(landmark_group, n_expected_points, group):
-    r"""
-    Ensure that the input matches the number of expected points.
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        Landmark group to validate
-    n_expected_points : `int`
-        Number of expected points
-    group : `str`
-        Group label for error message
-
-    Raises
-    ------
-    LabellingError
-        If the number of expected points doesn't match the number of given
-        points
-    """
-    n_points = landmark_group.lms.n_points
-    if n_points != n_expected_points:
-        raise LabellingError(_build_labelling_error_msg(group,
-                                                        n_expected_points,
-                                                        n_points))
-
-
-def _relabel_group_from_dict(pointcloud, labels_to_ranges):
-    """
-    Label the given pointcloud according to the given ordered dictionary
-    of labels to ranges. This assumes that you can semantically label the group
-    by using ranges in to the existing points e.g ::
-
-        labels_to_ranges = {'jaw': (0, 17, False)}
-
-    The third element of the range tuple is whether the range is a closed loop
-    or not. For example, for an eye landmark this would be ``True``, as you
-    do want to create a closed loop for an eye.
-
-    Parameters
-    ----------
-    pointcloud : :map:`PointCloud`
-        The pointcloud to apply semantic labels to.
-    labels_to_ranges : `OrderedDict`
-        Ordered dictionary of string labels to range tuples.
-
-    Returns
-    -------
-    landmark_group: :map:`LandmarkGroup`
-        New landmark group
-
-    Raises
-    ------
-    :class:`menpo.landmark.exceptions.LabellingError`
-        If the given pointcloud contains less than ``n_expected_points``
-        points.
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    n_points = pointcloud.n_points
-    masks = OrderedDict()
-    adjacency_lists = []
-    for label, tup in labels_to_ranges.items():
-        range_tuple = tup[:-1]
-        close_loop = tup[-1]
-        adjacency_lists.append(_connectivity_from_range(
-            range_tuple, close_loop=close_loop))
-        masks[label] = _mask_from_range(range_tuple, n_points)
-    adjacency_array = np.vstack(adjacency_lists)
-
-    new_landmark_group = LandmarkGroup(
-        PointUndirectedGraph.init_from_edges(pointcloud.points,
-                                             adjacency_array), masks)
-
-    return new_landmark_group
+from menpo.landmark.labels.base import (
+    _validate_input, _connectivity_from_array, _relabel_group_from_dict,
+    _connectivity_from_range)
 
 
 def ibug_face_68(landmark_group):
@@ -1355,1095 +1226,3 @@ def ibug_tongue(landmark_group):
         ('bisector', (13, 19, False))
     ])
     return group, _relabel_group_from_dict(landmark_group.lms, labels)
-
-
-def ibug_hand(landmark_group):
-    """
-    Apply the ibug's "standard" 39 point semantic labels to the landmark group.
-
-    The group label will be ``ibug_hand``.
-
-    The semantic labels applied are as follows:
-
-      - thumb
-      - index
-      - middle
-      - ring
-      - pinky
-      - palm
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``ibug_hand``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 39 points
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'ibug_hand'
-    _validate_input(landmark_group, 39, group)
-
-    thumb_indices = np.arange(0, 5)
-    index_indices = np.arange(5, 12)
-    middle_indices = np.arange(12, 19)
-    ring_indices = np.arange(19, 26)
-    pinky_indices = np.arange(26, 33)
-    palm_indices = np.hstack((np.array([32, 25, 18, 11, 33, 34, 4]),
-                              np.arange(35, 39)))
-
-    thumb_connectivity = _connectivity_from_array(thumb_indices,
-                                                  close_loop=False)
-    index_connectivity = _connectivity_from_array(index_indices,
-                                                  close_loop=False)
-    middle_connectivity = _connectivity_from_array(middle_indices,
-                                                   close_loop=False)
-    ring_connectivity = _connectivity_from_array(ring_indices,
-                                                 close_loop=False)
-    pinky_connectivity = _connectivity_from_array(pinky_indices,
-                                                  close_loop=False)
-    palm_connectivity = _connectivity_from_array(palm_indices,
-                                                 close_loop=True)
-
-    total_conn = np.vstack((thumb_connectivity, index_connectivity,
-                            middle_connectivity, ring_connectivity,
-                            pinky_connectivity, palm_connectivity))
-
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points,
-                                             total_conn))
-
-    new_landmark_group['thumb'] = thumb_indices
-    new_landmark_group['index'] = index_indices
-    new_landmark_group['middle'] = middle_indices
-    new_landmark_group['ring'] = ring_indices
-    new_landmark_group['pinky'] = pinky_indices
-    new_landmark_group['palm'] = palm_indices
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def stickmen_pose(landmark_group):
-    """
-    Apply the stickmen "standard" 12 point semantic labels to the landmark
-    group.
-
-    The group label will be ``stickmen_pose``.
-
-    The semantic labels applied are as follows:
-
-      - torso
-      - right_upper_arm
-      - left_upper_arm
-      - right_lower_arm
-      - left_lower_arm
-      - head
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``stickmen_pose``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 12 points
-
-    References
-    ----------
-    .. [1] http://www.robots.ox.ac.uk/~vgg/data/stickmen/
-    """
-    group = 'stickmen_pose'
-    _validate_input(landmark_group, 12, group)
-    labels = OrderedDict([
-        ('torso', (0, 2, False)),
-        ('right_upper arm', (2, 4, False)),
-        ('left_upper arm', (4, 6, False)),
-        ('right_lower_arm', (6, 8, False)),
-        ('left_lower_arm', (8, 10, False)),
-        ('head', (10, 12, False))
-    ])
-    return group, _relabel_group_from_dict(landmark_group.lms, labels)
-
-
-def lsp_pose(landmark_group):
-    """
-    Apply the lsp "standard" 14 point semantic labels to the landmark
-    group.
-
-    The group label will be ``lsp_pose``.
-
-    The semantic labels applied are as follows:
-
-      - left_leg
-      - right_leg
-      - left_arm
-      - right_arm
-      - head
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``lsp_pose``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 14 points
-
-    References
-    ----------
-    .. [1] http://www.comp.leeds.ac.uk/mat4saj/lsp.html
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'lsp_pose'
-    _validate_input(landmark_group, 14, group)
-
-    left_leg_indices = np.arange(0, 3)
-    right_leg_indices = np.arange(3, 6)
-    left_arm_indices = np.arange(6, 9)
-    right_arm_indices = np.arange(9, 12)
-    head_indices = np.arange(12, 14)
-
-    left_leg_connectivity = _connectivity_from_array(left_leg_indices)
-    right_leg_connectivity = _connectivity_from_array(right_leg_indices)
-    left_arm_connectivity = _connectivity_from_array(left_arm_indices)
-    right_arm_connectivity = _connectivity_from_array(right_arm_indices)
-    head_connectivity = _connectivity_from_array(head_indices)
-
-    total_conn = np.vstack([left_leg_connectivity,
-                            right_leg_connectivity,
-                            left_arm_connectivity,
-                            right_arm_connectivity,
-                            head_connectivity])
-
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points,
-                                             total_conn))
-
-    new_landmark_group['left_leg'] = left_leg_indices
-    new_landmark_group['right_leg'] = right_leg_indices
-    new_landmark_group['left_arm'] = left_arm_indices
-    new_landmark_group['right_arm'] = right_arm_indices
-    new_landmark_group['head'] = head_indices
-
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def flic_pose(landmark_group):
-    """
-    Apply the flic "standard" 11 point semantic labels to the landmark
-    group.
-
-    The group label will be ``flic_pose``.
-
-    The semantic labels applied are as follows:
-
-      - left_arm
-      - right_arm
-      - hips
-      - face
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``flic_pose``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 11 points
-
-    References
-    ----------
-    .. [1] http://vision.grasp.upenn.edu/cgi-bin/index.php?n=VideoLearning.FLIC
-    """
-    group = 'flic_pose'
-    _validate_input(landmark_group, 11, group)
-    labels = OrderedDict([
-        ('left_arm', (0, 3, False)),
-        ('right_arm', (3, 6, False)),
-        ('hips', (6, 8, False)),
-        ('face', (8, 11, True))])
-
-    return group, _relabel_group_from_dict(landmark_group.lms, labels)
-
-
-def human36M_pose_32(landmark_group):
-    """
-    Apply the human3.6M "standard" 32 point semantic labels to the landmark
-    group.
-
-    The group label will be ``human36M_pose_32``.
-
-    The semantic labels applied are as follows:
-
-      - pelvis
-      - right_leg
-      - left_leg
-      - spine
-      - head
-      - left_arm
-      - left_hand
-      - right_arm
-      - right_hand
-      - torso
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``human36M_pose_32``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 32 points
-
-    References
-    ----------
-    .. [1] http://vision.imar.ro/human3.6m/
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'human36M_pose_32'
-    _validate_input(landmark_group, 32, group)
-
-    pelvis_indices = np.array([1, 0, 6])
-    right_leg_indices = np.array(range(1, 6))
-    left_leg_indices = np.array(range(6, 11))
-    spine_indices = np.array([11, 12, 13])
-    head_indices = np.array([13, 14, 15])
-    left_arm_indices = np.array([16, 17, 18, 19, 23])
-    left_hand_indices = np.array([20, 21, 22])
-    right_arm_indices = np.array([24, 25, 26, 27, 29, 31])
-    right_hand_indices = np.array([28, 29, 30])
-    torso_indices = np.array([0, 1, 25, 13, 17, 6])
-
-    pelvis_connectivity = _connectivity_from_array(pelvis_indices)
-    right_leg_connectivity = _connectivity_from_array(right_leg_indices)
-    left_leg_connectivity = _connectivity_from_array(left_leg_indices)
-    spine_connectivity = _connectivity_from_array(spine_indices)
-    head_connectivity = _connectivity_from_array(head_indices)
-    left_arm_connectivity = _connectivity_from_array(left_arm_indices)
-    left_hand_connectivity = _connectivity_from_array(left_hand_indices)
-    right_arm_connectivity = _connectivity_from_array(right_arm_indices)
-    right_hand_connectivity = _connectivity_from_array(right_hand_indices)
-    torso_connectivity = _connectivity_from_array(torso_indices,
-                                                  close_loop=True)
-
-    total_conn = np.vstack([
-        pelvis_connectivity, right_leg_connectivity, left_leg_connectivity,
-        spine_connectivity, head_connectivity, left_arm_connectivity,
-        left_hand_connectivity, right_arm_connectivity,
-        right_hand_connectivity, torso_connectivity])
-
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points,
-                                             total_conn))
-
-    new_landmark_group['pelvis'] = pelvis_indices
-    new_landmark_group['right_leg'] = right_leg_indices
-    new_landmark_group['left_leg'] = left_leg_indices
-    new_landmark_group['spine'] = spine_indices
-    new_landmark_group['head'] = head_indices
-    new_landmark_group['left_arm'] = left_arm_indices
-    new_landmark_group['left_hand'] = left_hand_indices
-    new_landmark_group['right_arm'] = right_arm_indices
-    new_landmark_group['right_hand'] = right_hand_indices
-    new_landmark_group['torso'] = torso_indices
-
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def human36M_pose_17(landmark_group):
-    """
-    Apply the human3.6M "standard" 17 point semantic labels (based on the
-    original semantic labels of Human3.6 but removing the annotations
-    corresponding to duplicate points, soles and palms) to the landmark group.
-
-    The group label will be ``human36M_pose_17``.
-
-    The semantic labels applied are as follows:
-
-      - pelvis
-      - right_leg
-      - left_leg
-      - spine
-      - head
-      - left_arm
-      - right_arm
-      - torso
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``human36M_pose_17``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 32 points
-
-    References
-    ----------
-    .. [1] http://vision.imar.ro/human3.6m/
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'human36M_pose_17'
-    _validate_input(landmark_group, 32, group)
-
-    pelvis_indices = np.array([1, 0, 4])
-    right_leg_indices = np.arange(1, 4)
-    left_leg_indices = np.arange(4, 7)
-    spine_indices = np.array([0, 7, 8])
-    head_indices = np.array([8, 9, 10])
-    left_arm_indices = np.array([8, 11, 12, 13])
-    right_arm_indices = np.array([8, 14, 15, 16])
-    torso_indices = np.array([0, 1, 14, 8, 11, 4])
-
-    pelvis_connectivity = _connectivity_from_array(pelvis_indices)
-    right_leg_connectivity = _connectivity_from_array(right_leg_indices)
-    left_leg_connectivity = _connectivity_from_array(left_leg_indices)
-    spine_connectivity = _connectivity_from_array(spine_indices)
-    head_connectivity = _connectivity_from_array(head_indices)
-    left_arm_connectivity = _connectivity_from_array(left_arm_indices)
-    right_arm_connectivity = _connectivity_from_array(right_arm_indices)
-    torso_connectivity = _connectivity_from_array(torso_indices,
-                                                  close_loop=True)
-
-    total_conn = np.vstack([
-        pelvis_connectivity, right_leg_connectivity, left_leg_connectivity,
-        spine_connectivity, head_connectivity, left_arm_connectivity,
-        right_arm_connectivity, torso_connectivity])
-
-    # Ignore duplicate points, sole and palms
-    ind = np.hstack((np.arange(0, 4), np.arange(6, 9), np.arange(12, 16),
-                     np.arange(17, 20), np.arange(25, 28)))
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points[ind],
-                                             total_conn))
-
-    new_landmark_group['pelvis'] = pelvis_indices
-    new_landmark_group['right_leg'] = right_leg_indices
-    new_landmark_group['left_leg'] = left_leg_indices
-    new_landmark_group['spine'] = spine_indices
-    new_landmark_group['head'] = head_indices
-    new_landmark_group['left_arm'] = left_arm_indices
-    new_landmark_group['right_arm'] = right_arm_indices
-    new_landmark_group['torso'] = torso_indices
-
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def streetscene_car_view_0(landmark_group):
-    """
-    Apply the 8 point semantic labels of the view 0  of the MIT Street Scene
-    Car dataset to the landmark group.
-
-    The group label will be ``streetscene_car_view_0``.
-
-    The semantic labels applied are as follows:
-
-      - front
-      - bonnet
-      - windshield
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``streetscene_car_view_0``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 20 points
-
-    References
-    ----------
-    .. [1] http://www.cs.cmu.edu/~vboddeti/alignment.html
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'streetscene_car_view_0'
-    _validate_input(landmark_group, 20, group)
-
-    front_indices = np.array([0, 1, 3, 2])
-    bonnet_indices = np.array([2, 3, 5, 4])
-    windshield_indices = np.array([4, 5, 7, 6])
-
-    front_connectivity = _connectivity_from_array(front_indices,
-                                                  close_loop=True)
-    bonnet_connectivity = _connectivity_from_array(bonnet_indices,
-                                                   close_loop=True)
-    windshield_connectivity = _connectivity_from_array(windshield_indices,
-                                                       close_loop=True)
-
-    total_conn = np.vstack((front_connectivity, bonnet_connectivity,
-                            windshield_connectivity))
-
-    ind = np.arange(8)
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points[ind],
-                                             total_conn))
-
-    new_landmark_group['front'] = front_indices
-    new_landmark_group['bonnet'] = bonnet_indices
-    new_landmark_group['windshield'] = windshield_indices
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def streetscene_car_view_1(landmark_group):
-    """
-    Apply the 14 point semantic labels of the view 1  of the MIT Street Scene
-    Car dataset to the landmark group.
-
-    The group label will be ``streetscene_car_view_1``.
-
-    The semantic labels applied are as follows:
-
-      - front
-      - bonnet
-      - windshield
-      - left_side
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``streetscene_car_view_1``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 20 points
-
-    References
-    ----------
-    .. [1] http://www.cs.cmu.edu/~vboddeti/alignment.html
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'streetscene_car_view_1'
-    _validate_input(landmark_group, 20, group)
-
-    front_indices = np.array([0, 1, 3, 2])
-    bonnet_indices = np.array([2, 3, 5, 4])
-    windshield_indices = np.array([4, 5, 7, 6])
-    left_side_indices = np.array([0, 2, 4, 6, 8, 9, 10, 11, 13, 12])
-
-    front_connectivity = _connectivity_from_array(front_indices,
-                                                  close_loop=True)
-    bonnet_connectivity = _connectivity_from_array(bonnet_indices,
-                                                   close_loop=True)
-    windshield_connectivity = _connectivity_from_array(windshield_indices,
-                                                       close_loop=True)
-    left_side_connectivity = _connectivity_from_array(left_side_indices,
-                                                      close_loop=True)
-
-    total_conn = np.vstack((front_connectivity, bonnet_connectivity,
-                            windshield_connectivity, left_side_connectivity))
-
-    ind = np.hstack((np.arange(9), np.array([10, 12, 14, 16, 18])))
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points[ind],
-                                             total_conn))
-
-    new_landmark_group['front'] = front_indices
-    new_landmark_group['bonnet'] = bonnet_indices
-    new_landmark_group['windshield'] = windshield_indices
-    new_landmark_group['left_side'] = left_side_indices
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def streetscene_car_view_2(landmark_group):
-    """
-    Apply the 10 point semantic labels of the view 2  of the MIT Street Scene
-    Car dataset to the landmark group.
-
-    The group label will be ``streetscene_car_view_2``.
-
-    The semantic labels applied are as follows:
-
-      - left_side
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: 'streetscene_car_view_2'
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 20 points
-
-    References
-    ----------
-    .. [1] http://www.cs.cmu.edu/~vboddeti/alignment.html
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'streetscene_car_view_2'
-    _validate_input(landmark_group, 20, group)
-
-    left_side_indices = np.array([0, 1, 2, 3, 4, 5, 6, 7, 9, 8])
-
-    left_side_connectivity = _connectivity_from_array(left_side_indices,
-                                                      close_loop=True)
-
-    total_conn = left_side_connectivity
-
-    ind = np.array([0, 2, 4, 6, 8, 10, 12, 14, 16, 18])
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points[ind],
-                                             total_conn))
-
-    new_landmark_group['left_side'] = left_side_indices
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def streetscene_car_view_3(landmark_group):
-    """
-    Apply the 14 point semantic labels of the view 3  of the MIT Street Scene
-    Car dataset to the landmark group.
-
-    The group label will be ``streetscene_car_view_2``.
-
-    The semantic labels applied are as follows:
-
-      - left_side
-      - rear windshield
-      - trunk
-      - rear
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``streetscene_car_view_3``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 20 points
-
-    References
-    ----------
-    .. [1] http://www.cs.cmu.edu/~vboddeti/alignment.html
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'streetscene_car_view_3'
-    _validate_input(landmark_group, 20, group)
-
-    left_side_indices = np.array([0, 1, 2, 3, 4, 6, 8, 10, 13, 12])
-    rear_windshield_indices = np.array([4, 5, 7, 6])
-    trunk_indices = np.array([6, 7, 9, 8])
-    rear_indices = np.array([8, 9, 11, 10])
-
-    left_side_connectivity = _connectivity_from_array(left_side_indices,
-                                                      close_loop=True)
-    rear_windshield_connectivity = _connectivity_from_array(
-        rear_windshield_indices, close_loop=True)
-    trunk_connectivity = _connectivity_from_array(trunk_indices,
-                                                  close_loop=True)
-    rear_connectivity = _connectivity_from_array(rear_indices, close_loop=True)
-
-    total_conn = np.vstack((left_side_connectivity,
-                            rear_windshield_connectivity,
-                            trunk_connectivity, rear_connectivity))
-
-    ind = np.array([0, 2, 4, 6, 8, 9, 10, 11, 12, 13, 14, 15, 16, 18])
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points[ind],
-                                             total_conn))
-
-    new_landmark_group['left_side'] = left_side_indices
-    new_landmark_group['rear_windshield'] = rear_windshield_indices
-    new_landmark_group['trunk'] = trunk_indices
-    new_landmark_group['rear'] = rear_indices
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def streetscene_car_view_4(landmark_group):
-    """
-    Apply the 14 point semantic labels of the view 4  of the MIT Street Scene
-    Car dataset to the landmark group.
-
-    The group label will be ``streetscene_car_view_4``.
-
-    The semantic labels applied are as follows:
-
-      - front
-      - bonnet
-      - windshield
-      - right_side
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: 'streetscene_car_view_4'
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 20 points
-
-    References
-    ----------
-    .. [1] http://www.cs.cmu.edu/~vboddeti/alignment.html
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'streetscene_car_view_4'
-    _validate_input(landmark_group, 20, group)
-
-    front_indices = np.array([0, 1, 3, 2])
-    bonnet_indices = np.array([2, 3, 5, 4])
-    windshield_indices = np.array([4, 5, 7, 6])
-    right_side_indices = np.array([8, 9, 10, 11, 13, 12, 1, 3, 5, 7])
-
-    front_connectivity = _connectivity_from_array(front_indices,
-                                                  close_loop=True)
-    bonnet_connectivity = _connectivity_from_array(bonnet_indices,
-                                                   close_loop=True)
-    windshield_connectivity = _connectivity_from_array(windshield_indices,
-                                                       close_loop=True)
-    right_side_connectivity = _connectivity_from_array(right_side_indices,
-                                                       close_loop=True)
-
-    total_conn = np.vstack((front_connectivity, bonnet_connectivity,
-                            windshield_connectivity,
-                            right_side_connectivity))
-
-    ind = np.hstack((np.arange(8), np.array([9, 11, 13, 15, 17, 19])))
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points[ind],
-                                             total_conn))
-
-    new_landmark_group['front'] = front_indices
-    new_landmark_group['bonnet'] = bonnet_indices
-    new_landmark_group['windshield'] = windshield_indices
-    new_landmark_group['right_side'] = right_side_indices
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def streetscene_car_view_5(landmark_group):
-    """
-    Apply the 10 point semantic labels of the view 5 of the MIT Street Scene
-    Car dataset to the landmark group.
-
-    The group label will be ``streetscene_car_view_5``.
-
-    The semantic labels applied are as follows:
-
-      - right_side
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``streetscene_car_view_5``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 20 points
-
-    References
-    ----------
-    .. [1] http://www.cs.cmu.edu/~vboddeti/alignment.html
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'streetscene_car_view_5'
-    _validate_input(landmark_group, 20, group)
-
-    right_side_indices = np.array([0, 1, 2, 3, 4, 5, 6, 7, 9, 8])
-
-    right_side_connectivity = _connectivity_from_array(right_side_indices,
-                                                       close_loop=True)
-
-    total_conn = right_side_connectivity
-
-    ind = np.array([1, 3, 5, 7, 9, 11, 13, 15, 17, 19])
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points[ind],
-                                             total_conn))
-
-    new_landmark_group['right_side'] = right_side_indices
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def streetscene_car_view_6(landmark_group):
-    """
-    Apply the 14 point semantic labels of the view 6  of the MIT Street Scene
-    Car dataset to the landmark group.
-
-    The group label will be ``streetscene_car_view_6``.
-
-    The semantic labels applied are as follows:
-
-      - right_side
-      - rear_windshield
-      - trunk
-      - rear
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``streetscene_car_view_3``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 20 points
-
-    References
-    ----------
-    .. [1] http://www.cs.cmu.edu/~vboddeti/alignment.html
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'streetscene_car_view_6'
-    _validate_input(landmark_group, 20, group)
-
-    right_side_indices = np.array([0, 1, 2, 3, 5, 7, 9, 11, 13, 12])
-    rear_windshield_indices = np.array([4, 5, 7, 6])
-    trunk_indices = np.array([6, 7, 9, 8])
-    rear_indices = np.array([8, 9, 11, 10])
-
-    right_side_connectivity = _connectivity_from_array(right_side_indices,
-                                                       close_loop=True)
-    rear_windshield_connectivity = _connectivity_from_array(
-        rear_windshield_indices, close_loop=True)
-    trunk_connectivity = _connectivity_from_array(trunk_indices,
-                                                  close_loop=True)
-    rear_connectivity = _connectivity_from_array(rear_indices, close_loop=True)
-
-    total_conn = np.vstack((right_side_connectivity,
-                            rear_windshield_connectivity,
-                            trunk_connectivity, rear_connectivity))
-
-    ind = np.array([1, 3, 5, 7, 8, 9, 10, 11, 12, 13, 14, 15, 17, 19])
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points[ind],
-                                             total_conn))
-
-    new_landmark_group['right_side'] = right_side_indices
-    new_landmark_group['rear_windshield'] = rear_windshield_indices
-    new_landmark_group['trunk'] = trunk_indices
-    new_landmark_group['rear'] = rear_indices
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def streetscene_car_view_7(landmark_group):
-    """
-    Apply the 8 point semantic labels of the view 0  of the MIT Street Scene
-    Car dataset to the landmark group.
-
-    The group label will be ``streetscene_car_view_7``.
-
-    The semantic labels applied are as follows:
-
-      - rear_windshield
-      - trunk
-      - rear
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``streetscene_car_view_7``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    error : :map:`LabellingError`
-        If the given landmark group contains less than 20 points
-
-    References
-    ----------
-    .. [1] http://www.cs.cmu.edu/~vboddeti/alignment.html
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'streetscene_car_view_7'
-    _validate_input(landmark_group, 20, group)
-
-    rear_windshield_indices = np.array([0, 1, 3, 2])
-    trunk_indices = np.array([2, 3, 5, 4])
-    rear_indices = np.array([4, 5, 7, 6])
-
-    rear_windshield_connectivity = _connectivity_from_array(
-        rear_windshield_indices, close_loop=True)
-    trunk_connectivity = _connectivity_from_array(trunk_indices,
-                                                  close_loop=True)
-    rear_connectivity = _connectivity_from_array(rear_indices, close_loop=True)
-
-    total_conn = np.vstack((rear_windshield_connectivity,
-                            trunk_connectivity, rear_connectivity))
-
-    ind = np.arange(8, 16)
-    new_landmark_group = LandmarkGroup.init_with_all_label(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points[ind],
-                                             total_conn))
-
-    new_landmark_group['rear_windshield'] = rear_windshield_indices
-    new_landmark_group['trunk'] = trunk_indices
-    new_landmark_group['rear'] = rear_indices
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def bu3dfe_83(landmark_group):
-    """
-    Apply the BU-3DFE (Binghamton University 3D Facial Expression)
-    Database 83 point facial annotation markup to this landmark group.
-
-
-    The group label will be ``bu3dfe_83``.
-
-    The semantic labels applied are as follows:
-
-      - right_eye
-      - left_eye
-      - right_eyebrow
-      - left_eyebrow
-      - right_nose
-      - left_nose
-      - nostrils
-      - outer_mouth
-      - inner_mouth
-      - jaw
-
-    Parameters
-    ----------
-    landmark_group : :map:`LandmarkGroup`
-        The landmark group to apply semantic labels to.
-
-    Returns
-    -------
-    group : `str`
-        The group label: ``bu3dfe_83``
-    landmark_group : :map:`LandmarkGroup`
-        New landmark group.
-
-    Raises
-    ------
-    :class:`menpo.landmark.exceptions.LabellingError`
-        If the given landmark group contains less than 83 points
-
-    References
-    ----------
-    .. [1] http://www.cs.binghamton.edu/~lijun/Research/3DFE/3DFE_Analysis.html
-    """
-    from menpo.shape import PointUndirectedGraph
-
-    group = 'bu3dfe_83'
-    n_points = 83
-    _validate_input(landmark_group, n_points, group)
-
-    reye_indices = np.arange(0, 8)
-    leye_indices = np.arange(8, 16)
-    rbrow_indices = np.arange(16, 26)
-    lbrow_indices = np.arange(26, 36)
-    rnose_indices = np.arange(36, 39)
-    lnose_indices = np.arange(39, 42)
-    nostril_indices = np.arange(42, 48)
-    outermouth_indices = np.arange(48, 60)
-    innermouth_indices = np.arange(60, 68)
-    jaw_indices = np.arange(68, 83)
-
-    reye_connectivity = _connectivity_from_array(reye_indices, close_loop=True)
-    leye_connectivity = _connectivity_from_array(leye_indices, close_loop=True)
-    rbrow_connectivity = _connectivity_from_array(rbrow_indices,
-                                                  close_loop=True)
-    lbrow_connectivity = _connectivity_from_array(lbrow_indices,
-                                                  close_loop=True)
-    rnose_connectivity = _connectivity_from_array(rnose_indices)
-    nostril_connectivity = _connectivity_from_array(nostril_indices)
-    lnose_connectivity = _connectivity_from_array(lnose_indices)
-    outermouth_connectivity = _connectivity_from_array(outermouth_indices,
-                                                       close_loop=True)
-    innermouth_connectivity = _connectivity_from_array(innermouth_indices,
-                                                       close_loop=True)
-    jaw_connectivity = _connectivity_from_array(jaw_indices)
-
-    total_conn = np.vstack([
-        reye_connectivity, leye_connectivity,
-        rbrow_connectivity, lbrow_connectivity,
-        rnose_connectivity, nostril_connectivity, lnose_connectivity,
-        outermouth_connectivity, innermouth_connectivity,
-        jaw_connectivity
-    ])
-
-    new_landmark_group = LandmarkGroup(
-        PointUndirectedGraph.init_from_edges(landmark_group.lms.points,
-                                             total_conn),
-        OrderedDict([('all', np.ones(n_points, dtype=np.bool))]))
-
-    new_landmark_group['right_eye'] = reye_indices
-    new_landmark_group['left_eye'] = leye_indices
-    new_landmark_group['right_eyebrow'] = rbrow_indices
-    new_landmark_group['left_eyebrow'] = lbrow_indices
-    new_landmark_group['right_nose'] = rnose_indices
-    new_landmark_group['left_nose'] = lnose_indices
-    new_landmark_group['nostrils'] = nostril_indices
-    new_landmark_group['outer_mouth'] = outermouth_indices
-    new_landmark_group['inner_mouth'] = innermouth_indices
-    new_landmark_group['jaw'] = jaw_indices
-
-    del new_landmark_group['all']  # Remove pointless all group
-
-    return group, new_landmark_group
-
-
-def labeller(landmarkable, group, label_func):
-    """
-    Re-label an existing landmark group on a :map:`Landmarkable` object with a
-    new label set.
-
-    Parameters
-    ----------
-    landmarkable: :map:`Landmarkable`
-        :map:`Landmarkable` that will have it's :map:`LandmarkManager`
-        augmented with a new :map:`LandmarkGroup`
-    group: `str`
-        The group label of the existing landmark group that should be
-        re-labelled. A copy of this group will be attached to it's landmark
-        manager with new labels. The group label of this new group and the
-        labels it will have is determined by ``label_func``
-    label_func: `func`  -> `(str, LandmarkGroup)`
-        A labelling function taken from this module, Takes as input a
-        :map:`LandmarkGroup` and returns a tuple of
-        (new group label, new LandmarkGroup with semantic labels applied).
-
-    Returns
-    -------
-    landmarkable : :map:`Landmarkable`
-        Augmented ``landmarkable`` (this is just for convenience,
-        the object will actually be modified in place)
-    """
-    new_group, lmark_group = label_func(landmarkable.landmarks[group])
-    landmarkable.landmarks[new_group] = lmark_group
-    return landmarkable
